refactor data management file, added deploy api keys, removed deprecated methods
