--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "G:\\Anaconda3\\envs\\travis\\python.exe"
-}
