+++ config_template.yml
@@ -1,15 +1,15 @@
 # instructions: save this file as config.yml and add in required secret/keys, but don't commit that file
 recommender:  # main options for the recommender, if any
-    secret:  # secret key
-backend: local # which backend data to use
+    secret:   # secret key
+backend: gcp # which backend data to use
 data_local:  # options used for running a 'local' version of the recommender
     folder_pc: travis_optimiser\test_data  #  PC
     folder_unix: travis_optimiser/test_data  # MAC
     data_file: locations_recommender.csv
 data_gcp:  # links to cloud store and other places
-    project:  # gcp project name
+    project:   # gcp project name
     bucket:  # bucket location
-    folder:   # folder
-    data_file:  # file to download
-    json_key:   # location of json service account api key file, currently hungy's bucket
-google_key:  # google maps api key
+    folder:  # folder
+    data_file:  #  file to download
+    json_key:  # location of json service account api key file, currently hungy's bucket
+google_key:  #  google maps api key
