+++ travis_optimiser/recommender_data.py
@@ -7,21 +7,25 @@ import os
 Module for Management of the backend data needed for the recommender
 """
 
-def get_df_loc(method='local'):
+
+def get_df_loc(cfg, method='local'):
     """
     obtains the recommender data as a dataframe, used for getting the current
     copy of the data
-    TODO Alternate options, e.g. getting from S3 or GCP cloud store
+    method: local or gcp
     returns: DataFrame of location data avaliable
     """
     if method == 'local':
         dfLoc = load_data_from_local()
-    # TODO Implement alternative options
+    elif method == 'gcp':
+        dfLoc = load_data_from_gcp_cloud(cfg)
+    else:
+        print(f'error - non supported method: {method}')
     return dfLoc
-
 
 def update_poi_data(update_data, method='local'):
     """
+    TODO WIP function
     Updates the existing data of POIs used by recommender, depending on the option used
     """
     if method == 'local':
@@ -38,9 +42,21 @@ def load_data_from_local():
     dfLoc = pd.read_csv(os.path.join(folder, outfile), encoding='UTF-8')
     return dfLoc
 
-def load_data_from_gcp_cloud():
-    # To be implemented, when moving data to the cloud
-    pass
+def load_data_from_gcp_cloud(cfg):
+    # TODO Remove hardcoding of data being read and paths
+    bucket_folder = 'travis_recommender/csv_data/'  # folderpaths are in unix
+    file = "locations_recommender.csv"
+    # blob = bucket.blob(bucket_folder + file)
+    # filename = blob.name.split('/')[-1]
+    # local_folder = "scripts"
+    # dl_path = os.path.join(local_folder, filename)
+    # blob.download_to_filename(dl_path)
+    # print(f'{filename} downloaded from bucket.')
+
+    dl_path = 'gs://'+ bucket_folder + file
+    dfLoc = pd.read_csv(dl_path, encoding='UTF-8')
+    return dfLoc
 
 def update_data_to_local():
     pass
+
